backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.5
pydantic-settings==2.7.0

# Environment variables
python-dotenv==1.0.0

# CORS
python-multipart==0.0.6

# HTTP Client for image downloads
httpx==0.26.0

# Core dependencies
pandas==2.2.0
matplotlib>=3.9.0

# YouTube API
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Cache
redis==5.0.1

# Export
openpyxl==3.1.2

# Optional: LLM support (uncomment if needed)
# openai==1.12.0
# anthropic==0.18.0
